Initial proof-of-concept.
Allows a scene to be selected (and optionally recalled) and allows lights to be
added an removed from the scene. When updating the current scene, the current
state of all lights added to scene is saved as their state for said scene (this
is how the Hue API works).

There is nothing much else. No error-handling, very little feedback in general
and no proper layout whatsoever. Also, the code is still a mess...