Separate render-logic from main Hue class
Bit of general clean-up, moved all code interacting with the DOM outside
of the Hue class (which interacts with the Hue API). Added in two custom
events ("hueReady" and "hueResult") to tie everything together.

The "hueResult" event is used to output the response received from the
Hue API on the page (to get some feedback on actions for now).

Turns out this "vanilla" JavaScript thing is a b*tch as IE11 is by no
means standards compliant. And since everyone is using jQuery nobody
seems to actually know how it is supposed to work... At least IE on
WP8.1 and IE11 seem to be behaving mostly similar. Event creation-logic
is using a deprecated method at the moment, but that's the only one that
works in both IE11 and Firefox.